+++ CashRegister.BL/Reducers/RandomReducer.cs
@@ -1,17 +1,17 @@
 using System.Collections.Generic;
 using CashRegister.BL.Objects;
 
-namespace CashRegister.BL.Services
+namespace CashRegister.BL.Reducers
 {
-    public class MinReducer : IReducer
+    public class RandomReducer : IReducer
     {
-        public MinReducer()
+        public RandomReducer()
         {
         }
 
         public Denomination Reduce(IList<Denomination> resultList)
         {
-            return resultList.MinBy(x => x.TotalCoins);
+            return resultList.Random();
         }
     }
 }
